added code to read and write from and to a file

+++ 0x15-file_io/0-read_textfile.c
@@ -28,7 +28,7 @@ ssize_t read_textfile(const char *filename, size_t letters)
 		if ((nb_written < 0) || (nb_written != nb_read))
 			return (0);
 		n++;
-		if (n == letters)
+		if (n == (ssize_t)letters)
 			break;
 		nb_read = read(fd_open, &c, 1);
 	}
